+++ qti_package_maker/assessment_items/item_bank.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 # Standard Library
 import re
 import inspect