+++ qti_package_maker/common/color_theory/hue_layout.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 """Hue layout helpers."""
 
 # Standard Library