+++ qti_package_maker/common/color_theory/legacy_color_wheel.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 # Standard Library
 import math
 import random