+++ qti_package_maker/common/color_theory/rgb_color_name_match.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 """Closest xkcd color name helpers using seaborn."""
 
 # Standard Library