+++ qti_package_maker/common/color_theory/wheel_specs.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 """Wheel mode specifications and viewing defaults."""
 
 # Standard Library