+++ qti_package_maker/package_interface.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 # Standard Library
 import inspect
 import re